feat(user-interviews): show interview link on response page
Extract InterviewLinkCopyButton into a shared component and render it on
the per-person response page, so researchers can grab the unique interview
link from a single-person view — most useful while the response is still
awaiting.

Generated-By: PostHog Code
Task-Id: a4e48656-14c9-4293-aab1-fefa8c142695
