+++ products/user_interviews/frontend/UserInterview.tsx
@@ -1,18 +1,18 @@
 import { useValues } from 'kea'
 
-import { IconArrowLeft, IconCheck, IconChevronRight, IconClock, IconCopy } from '@posthog/icons'
+import { IconArrowLeft, IconCheck, IconChevronRight, IconClock } from '@posthog/icons'
 import { LemonButton, LemonSkeleton, LemonTag, LemonWidget } from '@posthog/lemon-ui'
 
 import { NotFound } from 'lib/components/NotFound'
 import { LemonMarkdown } from 'lib/lemon-ui/LemonMarkdown'
 import { Link } from 'lib/lemon-ui/Link'
-import { copyToClipboard } from 'lib/utils/copyToClipboard'
 import { SceneExport } from 'scenes/sceneTypes'
 import { urls } from 'scenes/urls'
 
 import { SceneContent } from '~/layout/scenes/components/SceneContent'
 
 import type { UserInterviewTopicApi } from './generated/api.schemas'
+import { InterviewLinkCopyButton } from './InterviewLinkCopyButton'
 import { UserInterviewLogicProps, userInterviewLogic } from './userInterviewLogic'
 
 export const scene: SceneExport<UserInterviewLogicProps> = {
@@ -215,39 +215,6 @@ function DetailRow({ label, value }: { label: string; value: string }): JSX.Elem
             <span className="text-muted text-sm">{label}</span>
             <span className="text-sm font-medium">{value}</span>
         </div>
-    )
-}
-
-function InterviewLinkCopyButton({ identifier, topicId }: { identifier: string; topicId: string }): JSX.Element {
-    const { linkForIdentifier, linksLoading, linksLoadFailed } = useValues(userInterviewLogic({ id: topicId }))
-    const interviewUrl = linkForIdentifier(identifier)
-
-    const handleCopy = (e: React.MouseEvent): void => {
-        e.preventDefault()
-        e.stopPropagation()
-        if (!interviewUrl) {
-            return
-        }
-        void copyToClipboard(interviewUrl, 'interview link')
-    }
-
-    const disabledReason = interviewUrl
-        ? undefined
-        : linksLoadFailed
-          ? "Couldn't generate link — refresh to retry"
-          : linksLoading
-            ? 'Generating link…'
-            : 'No link available'
-
-    return (
-        <LemonButton
-            type="tertiary"
-            size="xsmall"
-            icon={<IconCopy />}
-            onClick={handleCopy}
-            disabledReason={disabledReason}
-            tooltip="Copy interview link"
-        />
     )
 }
